# src/main/webapp/restapi/ui/swagger.json
{
  "swagger": "2.0",
  "info" : {
    "description" : "<h1>Angular CRUD Webapp<\/h1>",
    "version" : "1.0",
    "title" : "Angular CRUD REST API"
  },
  "basePath" : "/api/",
  "schemes" : [],
  "tags" : [
    {
      "name" : "UserEndpoint"
      ,
      "description" : "Created by phil on 1/14/2018."
    }
  ],
  "definitions" : {
    "json_Address" : {
      "type" : "object",
      "title" : "Address",
          "properties" : {
            "addressId" : {
            "readOnly" : false,
"description" : "",
"type" : "number"
            },
            "street" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "city" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "state" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "zip" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            }
          },
      "example" : {
  "addressId" : 12345,
  "street" : "...",
  "city" : "...",
  "state" : "...",
  "zip" : "..."
},
      "description" : "2018-12-21"
    }
    ,
    "json_DAOBean" : {
      "type" : "object",
      "title" : "DAOBean",
          "properties" : {
            "id" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "createDt" : {
            "readOnly" : false,
"description" : "",
"type" : "number"
            },
            "updateDt" : {
            "readOnly" : false,
"description" : "",
"type" : "number"
            },
            "updateBy" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "deleteDt" : {
            "readOnly" : false,
"description" : "",
"type" : "number"
            }
          },
      "example" : {
  "id" : "...",
  "createDt" : 12345,
  "updateDt" : 12345,
  "updateBy" : "...",
  "deleteDt" : 12345
},
      "description" : "2018-12-21"
    }
    ,
    "json_Email" : {
      "type" : "object",
      "title" : "Email",
          "properties" : {
            "type" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "value" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "primary" : {
            "readOnly" : false,
"description" : "",
"type" : "boolean"
            }
          },
      "example" : {
  "type" : "...",
  "value" : "...",
  "primary" : true
},
      "description" : "2018-12-21"
    }
    ,
    "json_Phone" : {
      "type" : "object",
      "title" : "Phone",
          "properties" : {
            "phoneId" : {
            "readOnly" : false,
"description" : "",
"type" : "number"
            },
            "number" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "type" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "areaCode" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "extension" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "primary" : {
            "readOnly" : false,
"description" : "",
"type" : "boolean"
            }
          },
      "example" : {
  "phoneId" : 12345,
  "number" : "...",
  "type" : "...",
  "areaCode" : "...",
  "extension" : "...",
  "primary" : true
},
      "description" : "2018-12-21"
    }
    ,
    "json_User" : {
      "type" : "object",
      "title" : "User",
      "allOf" : [
        {
          "$ref" : "#/definitions/json_DAOBean"
        },
        {
          "properties" : {
            "userId" : {
            "readOnly" : false,
"description" : "",
"type" : "number"
            },
            "username" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "firstName" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "lastName" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "password" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "token" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "addresses" : {
            "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/definitions/json_Address"
}
            },
            "roles" : {
            "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "type" : "string"
}
            },
            "country" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "website" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            },
            "phones" : {
            "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/definitions/json_Phone"
}
            },
            "emails" : {
            "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/definitions/json_Email"
}
            },
            "showpassword" : {
            "readOnly" : false,
"description" : "",
"type" : "string"
            }
          }
        }
      ],
      "example" : {
  "userId" : 12345,
  "username" : "...",
  "firstName" : "...",
  "lastName" : "...",
  "password" : "...",
  "token" : "...",
  "addresses" : [ {
    "addressId" : 12345,
    "street" : "...",
    "city" : "...",
    "state" : "...",
    "zip" : "..."
  }, {
    "addressId" : 12345,
    "street" : "...",
    "city" : "...",
    "state" : "...",
    "zip" : "..."
  } ],
  "roles" : [ "...", "..." ],
  "country" : "...",
  "website" : "...",
  "phones" : [ {
    "phoneId" : 12345,
    "number" : "...",
    "type" : "...",
    "areaCode" : "...",
    "extension" : "...",
    "primary" : true
  }, {
    "phoneId" : 12345,
    "number" : "...",
    "type" : "...",
    "areaCode" : "...",
    "extension" : "...",
    "primary" : true
  } ],
  "emails" : [ {
    "type" : "...",
    "value" : "...",
    "primary" : true
  }, {
    "type" : "...",
    "value" : "...",
    "primary" : true
  } ],
  "showpassword" : "...",
  "id" : "...",
  "createDt" : 12345,
  "updateDt" : 12345,
  "updateBy" : "...",
  "deleteDt" : 12345
},
      "description" : "Created by phil on 1/14/2018."
    }
  },
  "paths": {
    "\/user" : {
      "get" : {
        "tags" : [ "UserEndpoint" ],
        "description" : "",
        "operationId" : "getUsers",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "Authorization",
            "in" : "header",
            "type" : "string",
            "description" : ""
          }
        ],
        "responses" : {
          "200" : {
            "schema" : {
"description" : "",
"type" : "file"
            },
            "description" : ""
          }
        }
      }
      ,
      "post" : {
        "tags" : [ "UserEndpoint" ],
        "description" : "",
        "operationId" : "addUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "body",
            "in" : "body",
            "schema" : {
"$ref" : "#/definitions/json_User"
            },
            "description" : ""
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"description" : "",
"type" : "file"
            },
            "description" : ""
          }
        }
      }
      ,
      "put" : {
        "tags" : [ "UserEndpoint" ],
        "description" : "",
        "operationId" : "updateUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "body",
            "in" : "body",
            "schema" : {
"$ref" : "#/definitions/json_User"
            },
            "description" : ""
          }
        ],
        "responses" : {
          "204" : {
            "schema" : {
"description" : "",
"type" : "file"
            },
            "description" : ""
          }
        }
      }
    }
    ,
    "\/user/login" : {
      "post" : {
        "tags" : [ "UserEndpoint" ],
        "description" : "",
        "operationId" : "authenticate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "body",
            "in" : "body",
            "schema" : {
"description" : "",
"type" : "string"
            },
            "description" : ""
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"description" : "",
"type" : "file"
            },
            "description" : ""
          }
        }
      }
    }
    ,
    "\/user/{userId}" : {
      "delete" : {
        "tags" : [ "UserEndpoint" ],
        "description" : "",
        "operationId" : "deleteUser",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "userId",
            "in" : "path",
            "required" : true,
            "type" : "string",
            "description" : ""
          }
        ],
        "responses" : {
          "204" : {
            "schema" : {
"description" : "",
"type" : "file"
            },
            "description" : ""
          }
        }
      }
      ,
      "get" : {
        "tags" : [ "UserEndpoint" ],
        "description" : "",
        "operationId" : "getUser",
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "userId",
            "in" : "path",
            "required" : true,
            "type" : "string",
            "description" : ""
          }
        ],
        "responses" : {
          "200" : {
            "schema" : {
"description" : "",
"type" : "file"
            },
            "description" : ""
          }
        }
      }
    }
  }
}